Reorganise module layout to be more hermetic (#50)
It was strange that the shared utilities had to be installed as a
package, as they should be just an implementation detail. This commit
moves all the modules into a new `iree-kernel-benchmark` directory,
creating an implicit namespace package that can optionally be installed,
makes the command-line tools be invoked in module style, and switches to
using explicit relative imports for all internal dependencies.

Resolves #49.