gemmbench: Generate benchmarks from supplied dtypes instead of filtering (#76)
Prior to this commit, gemmbench's problems.py hardcoded a set of
datatypes for each benchmark, generating all possibilities ahead of
time. The --dtypes and --raw_accumulators command-line arguments were
then used to filter the resulting set of benchmarks.

This commit refactors problems.py to have less redundancy and to not
hardcode any datatypes. Now only the shapes are hardcoded, and the
--dtypes and --raw_accumulators command-line arguments control which
datatypes are used in the result. One consequence of this is that now we
get the same set of shapes no matter which datatype is requested, which
provides more thorough testing for types like i8 that previously only
had a very small number of hardcoded shapes.

To avoid an explosion in the amount of tests, the default set of
datatypes is changed to just f16. CI is changed to run f16 and i8 (in
separate job steps for better visibility); note that bf16 CI coverage is
removed.

Also, the padded LLaMA shapes are removed.