Update model benchmarks to matmul_transpose_b (#75)
In general, we try to use matmul_transpose_b (the tensors are M x K and
N x K in memory) for most matrux multiplications in IREE today. However,
most of the tests that hardcoded a layout were using matmul_transpose_a
instead, leading to insufficient benchmark coverage of the common case.